Linux module: Replace disable/enable macros with proper local_irq_*
The disable() and enable() macros were used to disable/enable
interrupts, by locking/unlocking a spinlock.  Not only did they refer to
local variables (a big no-no), but the spinlock is also unnecessary, as
it's an automatic variable and only exists in the function itself.

This replaces them all with proper local_irq_save/restore().